refactor: rename viewmodel files and update imports
