+++ lib/views/home.dart
@@ -1,6 +1,6 @@
 import 'package:twokey/views/keys/keys.dart';
 import 'package:twokey/views/settings.dart';
-import 'package:twokey/viewmodels/navigation_viewmodel.dart';
+import 'package:twokey/viewmodels/navigation.dart';
 import 'package:twokey/views/logger.dart';
 import 'package:twokey/widgets/adaptive_scaffold.dart';
 import 'package:flutter/material.dart';
