+++ lib/views/keys/keys.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:twokey/viewmodels/keys_viewmodel.dart';
+import 'package:twokey/viewmodels/keys.dart';
 import 'widgets/device_info_section.dart';
 import 'widgets/credentials_section.dart';
 import 'widgets/developer_tools_section.dart';
