+++ lib/views/keys/widgets/developer_tools_section.dart
@@ -1,7 +1,7 @@
 import 'package:twokey/widgets/card.dart';
 import 'package:twokey/common/context.dart';
 import 'package:flutter/material.dart';
-import 'package:twokey/viewmodels/keys_viewmodel.dart';
+import 'package:twokey/viewmodels/keys.dart';
 import 'package:provider/provider.dart';
 
 class DeveloperToolsSection extends StatelessWidget {
