+++ test/widget_test.dart
@@ -10,7 +10,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:provider/provider.dart';
 
 import 'package:twokey/main.dart';
-import 'package:twokey/viewmodels/navigation_viewmodel.dart';
+import 'package:twokey/viewmodels/navigation.dart';
 import 'package:twokey/widgets/adaptive_scaffold.dart';
 
 void main() {
